add: users router with tests

+++ madr_postgres/services/users.py
@@ -1,7 +1,9 @@
-from madr_postgres.exceptions.users import UserAlreadyExists
+from sqlalchemy.exc import IntegrityError
+
+from madr_postgres.exceptions.users import UserAlreadyExists, UserNotFound
 from madr_postgres.models.users import User
 from madr_postgres.repositories.users import UserRepository
-from madr_postgres.schemas.users import UserCreate
+from madr_postgres.schemas.users import UserCreate, UserUpdate
 from madr_postgres.utils.sanitize import sanitize_string
 
 
@@ -25,5 +27,22 @@ class UserService:
 
     async def delete_user(self, user_id: int):
         user = await self.repo.read_by_id(user_id)
+        if not user:
+            raise UserNotFound()
         await self.repo.delete(user)
         return {'message': 'User deleted'}
+
+    async def patch_user(self, user_id: int, user: UserUpdate):
+        db_user = await self.repo.read_by_id(user_id)
+
+        if not db_user:
+            raise UserNotFound()
+
+        for key, value in user.model_dump(exclude_unset=True).items():
+            setattr(db_user, key, value)
+        db_user.username = sanitize_string(db_user.username)
+
+        try:
+            return await self.repo.patch(db_user)
+        except IntegrityError:
+            raise UserAlreadyExists('Username or email')
